The default labeller converts options into labels
Fixes #3734
