+++ pkg/skaffold/config/options_test.go
@@ -23,103 +23,6 @@ import (
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
-func TestLabels(t *testing.T) {
-	tests := []struct {
-		description    string
-		options        SkaffoldOptions
-		expectedLabels map[string]string
-	}{
-		{
-			description:    "empty",
-			options:        SkaffoldOptions{},
-			expectedLabels: map[string]string{},
-		},
-		{
-			description: "cleanup",
-			options:     SkaffoldOptions{Cleanup: true},
-			expectedLabels: map[string]string{
-				"skaffold.dev/cleanup": "true",
-			},
-		},
-		{
-			description: "namespace",
-			options:     SkaffoldOptions{Namespace: "NS"},
-			expectedLabels: map[string]string{
-				"skaffold.dev/namespace": "NS",
-			},
-		},
-		{
-			description: "profile",
-			options:     SkaffoldOptions{Profiles: []string{"profile"}},
-			expectedLabels: map[string]string{
-				"skaffold.dev/profile.0": "profile",
-			},
-		},
-		{
-			description: "profiles",
-			options:     SkaffoldOptions{Profiles: []string{"profile1", "profile2"}},
-			expectedLabels: map[string]string{
-				"skaffold.dev/profile.0": "profile1",
-				"skaffold.dev/profile.1": "profile2",
-			},
-		},
-		{
-			description: "tail",
-			options:     SkaffoldOptions{Tail: true},
-			expectedLabels: map[string]string{
-				"skaffold.dev/tail": "true",
-			},
-		},
-		{
-			description: "tail dev",
-			options:     SkaffoldOptions{TailDev: true},
-			expectedLabels: map[string]string{
-				"skaffold.dev/tail": "true",
-			},
-		},
-		{
-			description: "all labels",
-			options: SkaffoldOptions{
-				Cleanup:   true,
-				Namespace: "namespace",
-				Profiles:  []string{"p1", "p2"},
-			},
-			expectedLabels: map[string]string{
-				"skaffold.dev/cleanup":   "true",
-				"skaffold.dev/namespace": "namespace",
-				"skaffold.dev/profile.0": "p1",
-				"skaffold.dev/profile.1": "p2",
-			},
-		},
-		{
-			description: "custom labels",
-			options: SkaffoldOptions{
-				Cleanup: true,
-				CustomLabels: []string{
-					"one=first",
-					"two=second",
-					"three=",
-					"four",
-				},
-			},
-			expectedLabels: map[string]string{
-				"skaffold.dev/cleanup": "true",
-				"one":                  "first",
-				"two":                  "second",
-				"three":                "",
-				"four":                 "",
-			},
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			labels := test.options.Labels()
-
-			t.CheckDeepEqual(test.expectedLabels, labels)
-		})
-	}
-}
-
 func TestIsTargetImage(t *testing.T) {
 	tests := []struct {
 		description   string
