+++ pkg/skaffold/deploy/labeller_test.go
@@ -19,6 +19,8 @@ package deploy
 import (
 	"testing"
 
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/config"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/version"
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
@@ -40,7 +42,13 @@ func TestDefaultLabeller(t *testing.T) {
 	}
 	for _, test := range tests {
 		testutil.Run(t, test.description, func(t *testutil.T) {
-			l := NewLabeller(test.version)
+			t.Override(&version.Get, func() *version.Info {
+				return &version.Info{
+					Version: test.version,
+				}
+			})
+
+			l := NewLabeller(config.SkaffoldOptions{})
 			labels := l.Labels()
 
 			expected := map[string]string{
@@ -53,10 +61,10 @@ func TestDefaultLabeller(t *testing.T) {
 }
 
 func TestDefaultLabeller_TwoInstancesHaveSameRunID(t *testing.T) {
-	first := NewLabeller("v1.0.0")
-	second := NewLabeller("v2.0.0")
+	first := NewLabeller(config.SkaffoldOptions{})
+	second := NewLabeller(config.SkaffoldOptions{})
 
-	if first.RunIDKeyValueString() != second.RunIDKeyValueString() {
+	if first.RunIDSelector() != second.RunIDSelector() {
 		t.Errorf("expected the run-id to be the same for two instances")
 	}
 	if first.runID == "" {
@@ -64,13 +72,115 @@ func TestDefaultLabeller_TwoInstancesHaveSameRunID(t *testing.T) {
 	}
 }
 
-func TestK8sManagedByLabelKeyValueString(t *testing.T) {
-	defaultLabeller := &DefaultLabeller{
-		version: "version",
+func TestDefaultLabeller_OverrideRunId(t *testing.T) {
+	labeller := NewLabeller(config.SkaffoldOptions{
+		CustomLabels: []string{RunIDLabel + "=ID"},
+	})
+
+	labels := labeller.Labels()
+
+	testutil.CheckDeepEqual(t, "ID", labels[RunIDLabel])
+}
+
+func TestLabels(t *testing.T) {
+	tests := []struct {
+		description    string
+		options        config.SkaffoldOptions
+		expectedLabels map[string]string
+	}{
+		{
+			description:    "empty",
+			options:        config.SkaffoldOptions{},
+			expectedLabels: map[string]string{},
+		},
+		{
+			description: "cleanup",
+			options:     config.SkaffoldOptions{Cleanup: true},
+			expectedLabels: map[string]string{
+				"skaffold.dev/cleanup": "true",
+			},
+		},
+		{
+			description: "namespace",
+			options:     config.SkaffoldOptions{Namespace: "NS"},
+			expectedLabels: map[string]string{
+				"skaffold.dev/namespace": "NS",
+			},
+		},
+		{
+			description: "profile",
+			options:     config.SkaffoldOptions{Profiles: []string{"profile"}},
+			expectedLabels: map[string]string{
+				"skaffold.dev/profile.0": "profile",
+			},
+		},
+		{
+			description: "profiles",
+			options:     config.SkaffoldOptions{Profiles: []string{"profile1", "profile2"}},
+			expectedLabels: map[string]string{
+				"skaffold.dev/profile.0": "profile1",
+				"skaffold.dev/profile.1": "profile2",
+			},
+		},
+		{
+			description: "tail",
+			options:     config.SkaffoldOptions{Tail: true},
+			expectedLabels: map[string]string{
+				"skaffold.dev/tail": "true",
+			},
+		},
+		{
+			description: "tail dev",
+			options:     config.SkaffoldOptions{TailDev: true},
+			expectedLabels: map[string]string{
+				"skaffold.dev/tail": "true",
+			},
+		},
+		{
+			description: "all labels",
+			options: config.SkaffoldOptions{
+				Cleanup:   true,
+				Namespace: "namespace",
+				Profiles:  []string{"p1", "p2"},
+			},
+			expectedLabels: map[string]string{
+				"skaffold.dev/cleanup":   "true",
+				"skaffold.dev/namespace": "namespace",
+				"skaffold.dev/profile.0": "p1",
+				"skaffold.dev/profile.1": "p2",
+			},
+		},
+		{
+			description: "custom labels",
+			options: config.SkaffoldOptions{
+				Cleanup: true,
+				CustomLabels: []string{
+					"one=first",
+					"two=second",
+					"three=",
+					"four",
+				},
+			},
+			expectedLabels: map[string]string{
+				"skaffold.dev/cleanup": "true",
+				"one":                  "first",
+				"two":                  "second",
+				"three":                "",
+				"four":                 "",
+			},
+		},
 	}
-	expected := "app.kubernetes.io/managed-by=skaffold-version"
-	actual := defaultLabeller.K8sManagedByLabelKeyValueString()
-	if actual != expected {
-		t.Fatalf("actual label not equal to expected label. Actual: \n %s \n Expected: \n %s", actual, expected)
+	for _, test := range tests {
+		testutil.Run(t, test.description, func(t *testutil.T) {
+			l := NewLabeller(test.options)
+
+			labels := l.Labels()
+
+			// Ignore those two labels for this test
+			delete(labels, K8sManagedByLabelKey)
+			delete(labels, RunIDLabel)
+
+			t.CheckDeepEqual(test.expectedLabels, labels)
+		})
 	}
 }
